Make Grafana-style HTML reporter the default and only option
Breaking changes:
- Replace html_reporter.py with Grafana-style version
- Remove --style argument from evaluate.py
- HTMLReporter now uses Grafana dark theme by default
- Keep html_reporter_classic.py as backup reference

Changes:
- Rename HTMLReporterGrafana class to HTMLReporter
- Remove style selection logic from evaluate.py
- Simplify report generation (no more classic vs grafana choice)

The Grafana-style reporter is now the standard HTML output with:
- Dark theme matching Grafana aesthetics
- Konveyor logo in header
- Comprehensive scoring algorithm
- Scalable performance charts (200-400 rules)
- Security issue details
- Top performers podium with all metrics
